Change log -	Lab answer key: Module 04: Lab A: Deploying Software-Defined Networking
o	Ex 1, Task 2: I verified that only the **DC** VM will require you to run `slmgr -rearm` and be restarted.  I did not remove the part of the note that says to do slmgr -rearm on all the VMs. This note can simply say “Sign in to the **DC** VM using the CORP\LabAdmin username and LS1setup! Password, run `slmgr -rearm` and restart it.”